+++ src/components/Faculty/FeedbackFrom.tsx
@@ -11,9 +11,11 @@ import {
 import TextArea from "antd/es/input/TextArea";
 import { Button } from "@/components/ui/button";
 import { useState, useEffect } from "react";
-import { updateApproval } from "@/helpers/api";
+import { updateApproval } from "@/helpers//faculty/api";
 import Cookies from "js-cookie";
 import { useRouter } from "next/navigation";
+import tick from "@/../public/tick.svg";
+import cross from "@/../public/cross.svg";
 
 interface ButtonProps {
   children?: React.ReactNode;
@@ -48,31 +50,7 @@ const AcceptButton: React.FC<ButtonProps> = ({
       }}
     >
       Approve
-      <svg
-        xmlns="http://www.w3.org/2000/svg"
-        x="0px"
-        y="0px"
-        width="2rem"
-        height="2rem"
-        viewBox="0 0 40 40"
-        className="ml-2"
-      >
-        <path
-          fill="#8bb7f0"
-          d="M4.769,37.5c-1.251,0-2.269-1.018-2.269-2.269V4.769C2.5,3.518,3.518,2.5,4.769,2.5h30.462 c1.251,0,2.269,1.018,2.269,2.269v30.462c0,1.251-1.018,2.269-2.269,2.269H4.769z"
-        ></path>
-        <path
-          fill="#4e7ab5"
-          d="M35.231,3C36.206,3,37,3.794,37,4.769v30.462C37,36.206,36.206,37,35.231,37H4.769 C3.794,37,3,36.206,3,35.231V4.769C3,3.794,3.794,3,4.769,3H35.231 M35.231,2H4.769C3.24,2,2,3.24,2,4.769v30.462 C2,36.76,3.24,38,4.769,38h30.462C36.76,38,38,36.76,38,35.231V4.769C38,3.24,36.76,2,35.231,2L35.231,2z"
-        ></path>
-        <path
-          fill="none"
-          stroke="#fff"
-          strokeMiterlimit="10"
-          strokeWidth="3"
-          d="M11 20.053L16.964 26.018 30.385 12.598"
-        ></path>
-      </svg>
+      <img src={tick.src} />
     </Button>
   );
 };
@@ -94,27 +72,7 @@ const RejectButton: React.FC<ButtonProps> = ({
       }}
     >
       Reject
-      <svg
-        xmlns="http://www.w3.org/2000/svg"
-        x="0px"
-        y="0px"
-        width="2rem"
-        height="2rem"
-        viewBox="0 0 48 48"
-      >
-        <path
-          fill="#f44336"
-          d="M44,24c0,11.045-8.955,20-20,20S4,35.045,4,24S12.955,4,24,4S44,12.955,44,24z"
-        ></path>
-        <path
-          fill="#fff"
-          d="M29.656,15.516l2.828,2.828l-14.14,14.14l-2.828-2.828L29.656,15.516z"
-        ></path>
-        <path
-          fill="#fff"
-          d="M32.484,29.656l-2.828,2.828l-14.14-14.14l2.828-2.828L32.484,29.656z"
-        ></path>
-      </svg>
+      <img src={cross.src} alt="" />
     </Button>
   );
 };
